requirements.txt
numpy
pandas
pyyaml
scikit-learn
joblib
matplotlib
